test: Add comprehensive unit tests for rx_bus manager and OneWire driver
Implements complete test coverage for the rx_bus abstraction layer:

- test_rx_bus_manager: 36 tests covering initialization, bus registration,
  removal, lookup, thread-safe callbacks, and command pattern

- test_rx_bus_onewire: 55 tests covering initialization, reset/presence
  detection, bit/byte/buffer operations, ROM commands, search algorithm,
  and GPIO error injection

Key features:
- Mock hardware infrastructure (CMT timer, system registers, GPIO, CRC)
- Auto-incrementing timer mock prevents infinite loops in delay functions
- Mock headers shadow real hardware headers for host-side testing
- All tests pass (91 total tests across 13 test suites)

Resolves #61

+++ star-rx72n-firmware/tests/mocks/mock_rx_crc.h
@@ -14,7 +14,7 @@
 #ifndef MOCK_RX_CRC_H
 #define MOCK_RX_CRC_H
 
-#include <stddef.h>
+#include <stdbool.h>
 #include <stdint.h>
 
 #ifdef __cplusplus
@@ -32,7 +32,7 @@ extern "C" {
  *
  * @return Computed CRC-8 value
  */
-uint8_t rx_crc8_maxim(const uint8_t* data, size_t length);
+uint8_t rx_crc8_maxim(const uint8_t* data, uint32_t length);
 
 /**
  * @brief Set fixed CRC-8 value to return (for testing)
